♻️ refactor: useServiceInitialization에 단일 책임 원칙 적용
- initializeServices를 작은 집중된 함수들로 분리 (createUIUpdater, createTimerService, createCartEventHandler)
- 각 함수가 명확한 하나의 책임만 가지도록 개선
- 코드 가독성과 유지보수성 향상
- CartEventHandler 생성자 호출을 한 줄로 단순화
- 동일한 기능을 유지하면서 코드 구조 개선
